Extract getting of the gcov version to structs
